requirements.txt
discord.py[voice]>=2.3.2
openai-whisper>=20231117
pyttsx3>=2.90
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0